requirements.txt
# BiRefNet Alpha Mask Toolkit Requirements
# Core dependencies for AI-powered human segmentation and alpha mask generation

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.9.0
scikit-image>=0.19.0

# Computer Vision Models
timm>=0.9.0
kornia>=0.6.8
einops>=0.6.0

# Progress and Utilities
tqdm>=4.64.0
prettytable>=3.5.0
tabulate>=0.9.0

# HuggingFace Integration
huggingface-hub>=0.16.0
accelerate>=0.20.0

# Optional: For CUDA acceleration (recommended)
# Install CUDA-compatible PyTorch from: https://pytorch.org/get-started/locally/

# System Requirements:
# - Python 3.10+
# - CUDA-capable GPU (recommended)
# - 8GB+ RAM
# - 2GB+ storage for model weights